% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ukbmisc-package.R
\docType{package}
\name{ukbmisc-package}
\alias{ukbmisc}
\alias{ukbmisc-package}
\title{ukbmisc: Link UKB datasets}
\description{
Miscellaneous utility functions for working with UK Biobank data.
}
\author{
\strong{Maintainer}: Alasdair Warwick \email{alasdair.warwick06@gmail.com}

}
\keyword{internal}
